feat(inspector): Split settings into two components
Split dialog button from dialog itself. Control the dialog from
global state.
