+++ inspector/web/components/settings-button.css
@@ -13,133 +13,3 @@ settings-button .menu-button {
 settings-button .menu-button:hover {
   background: #ddd;
 }
-
-settings-button dialog {
-  display: none;
-  flex-direction: column;
-  border: 1px solid var(--color-border);
-  padding: 0;
-  border-radius: 6px;
-  pointer-events: none;
-  width: 400px;
-  min-height: 200px;
-}
-
-settings-button dialog[open] {
-  display: flex;
-  pointer-events: auto;
-}
-
-settings-button .dialog-header,
-settings-button .dialog-content,
-settings-button .dialog-footer {
-  padding-block: var(--space-xs);
-  padding-inline: var(--space-md);
-}
-
-settings-button .dialog-header {
-  border-block-end: 1px solid var(--color-border);
-  position: relative;
-}
-
-settings-button dialog .close-dialog-button {
-  display: flex;
-  position: absolute;
-  inset-block-start: 2px;
-  inset-inline-end: 2px;
-  background-color: transparent;
-  padding: 0;
-  border: none;
-}
-
-settings-button dialog .close-dialog-button:hover {
-  background: #ddd;
-  border: solid 1x var(--color-border);
-  border-radius: 6px;
-}
-
-settings-button .dialog-content {
-  flex-grow: 1;
-}
-
-settings-button .dialog-footer {
-  border-block-start: 1px solid var(--color-border);
-  display: flex;
-  flex-direction: row-reverse;
-  gap: 6px;
-}
-
-settings-button .dialog-footer button {
-  display: flex;
-  cursor: pointer;
-  border: none;
-  color: #555;
-  border-radius: 2px;
-  padding-block: 6px;
-  padding-inline: 6px;
-  font-weight: 500;
-  border: 1px solid transparent;
-}
-
-settings-button .dialog-footer button:hover {
-  background: #ddd;
-}
-
-settings-button dialog h2 {
-  font-size: var(--text-lg);
-  margin-block: var(--space-md);
-  margin-inline-end: var(--space-md);
-}
-
-settings-button .settings-form,
-settings-button .read-only-settings {
-  padding-block: var(--space-sm);
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-sm);
-}
-
-settings-button label {
-  font-weight: 500;
-}
-
-settings-button input {
-  border: 1px solid var(--color-border);
-  background: var(--color-raised);
-  border-radius: var(--rounded-sm);
-  padding-block: var(--space-sm);
-  padding-inline: var(--space-sm);
-  width: 100%;
-}
-
-settings-button .value-row {
-  display: flex;
-  flex-direction: row;
-  gap: var(--space-xs);
-}
-
-settings-button .value {
-  border: 1px solid var(--color-border);
-  background: var(--color-raised);
-  border-radius: var(--rounded-sm);
-  padding-block: var(--space-sm);
-  padding-inline: var(--space-sm);
-  width: 100%;
-}
-
-settings-button .value-row button {
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  border: none;
-  color: #555;
-  border-radius: var(--rounded-sm);
-  padding-block: calc(var(--space-sm) - 1px);
-  padding-inline: var(--space-sm);
-  font-weight: 500;
-  border: 1px solid var(--color-border);
-}
-
-settings-button .saved-values button:hover {
-  background: #ddd;
-}
