+++ inspector/web/components/settings-button.ts
@@ -1,63 +1,16 @@
 import { html, LitElement } from 'lit';
-import { customElement, property } from 'lit/decorators.js';
-import { StorageData, store } from '../lib/store';
+import { customElement } from 'lit/decorators.js';
+import { store } from '../lib/store';
 import './settings-button.css';
 
 @customElement('settings-button')
 export class SettingsButton extends LitElement {
-  @property({ type: Boolean })
-  open: boolean = false;
-
-  @property({ type: String })
-  openAIAPIToken: string = '';
-
   createRenderRoot() {
     return this;
   }
 
-  connectedCallback() {
-    store.subscribe((data) => {
-      this.openAIAPIToken = data.settings?.openAIAPIToken;
-    });
-    super.connectedCallback();
-  }
-
-  onSave(e: SubmitEvent) {
-    e.preventDefault();
-    const form = e.target as HTMLFormElement;
-    const input = form.elements.namedItem('openAIAPIToken') as HTMLInputElement;
-    const openAIAPIToken = input.value;
-    const settings = store.get().settings;
-
-    store.update({ settings: { openAIAPIToken, ...settings } });
-
-    this.toggleDialog();
-  }
-
-  onClear(e: SubmitEvent) {
-    e.preventDefault();
-
-    const settings = store.get().settings;
-    delete settings.openAIAPIToken;
-
-    store.update({ settings: settings });
-  }
-
-  onDismiss() {
-    // 1. clear any data
-    // 2. close modal
-    this.toggleDialog();
-  }
-
-  toggleDialog() {
-    const dialog = this.querySelector<HTMLDialogElement>(
-      '[data-settings-dialog]'
-    );
-    if (dialog.open) {
-      dialog.close();
-    } else {
-      dialog.showModal();
-    }
+  openDialog() {
+    store.update({ settingsDialogOpen: true });
   }
 
   render() {
@@ -65,7 +18,7 @@ export class SettingsButton extends LitElement {
       <button
         aria-label="open settings"
         class="menu-button"
-        @click="${this.toggleDialog}"
+        @click="${this.openDialog}"
       >
         <svg
           focusable="false"
@@ -85,59 +38,6 @@ export class SettingsButton extends LitElement {
           <circle cx="12" cy="12" r="3" />
         </svg>
       </button>
-      <dialog data-settings-dialog>
-        <div class="dialog-header">
-          <h2>Settings</h2>
-          <button
-            aria-label="close dialog"
-            class="close-dialog-button"
-            @click="${this.onDismiss}"
-          >
-            <svg
-              focusable="false"
-              xmlns="http://www.w3.org/2000/svg"
-              width="24"
-              height="24"
-              viewBox="0 0 24 24"
-              fill="none"
-              stroke="currentColor"
-              stroke-width="1"
-              stroke-linecap="round"
-              stroke-linejoin="round"
-              class="lucide lucide-x"
-            >
-              <path d="M18 6 6 18" />
-              <path d="m6 6 12 12" />
-            </svg>
-          </button>
-        </div>
-        <div class="dialog-content">
-          ${this.openAIAPIToken
-            ? html`
-                <div class="read-only-settings">
-                  <label>OpenAI API token</label>
-                  <div class="value-row">
-                    <div class="value">••••••••••••••••</div>
-                    <button @click="${this.onClear}">Clear</button>
-                  </div>
-                </div>
-              `
-            : html`
-                <form
-                  id="settings-form"
-                  class="settings-form"
-                  @submit="${this.onSave}"
-                >
-                  <label for="openAIAPIToken">OpenAI API token</label>
-                  <input id="openAIAPIToken" name="openAIAPIToken" />
-                </form>
-              `}
-        </div>
-        <div class="dialog-footer">
-          <button type="submit" form="settings-form">Save</button>
-          <button @click="${this.onDismiss}">Cancel</button>
-        </div>
-      </dialog>
     `;
   }
 }
